Adapt FieldStorage to patched Python versions
The FieldStorage class now takes a separator as an additional parameter
in recent Python 3 versions as part of a security fix. This fix is not
contained in the last Python 2 version 2.7.18, but it has been backported
by some vendors as part of the extended support for Python 2.

Therefore, we need to take this into consideration as well, and also
change the splitting of query strings in POST requests so that this is
consistent with the behavior in GET requests.